Bug fixes
Fixed the Boss move speed so he doesn't move as fast in the build. Fixed the Pause menu in the boss level. Add code to the Quit buttons. Moved The platforms and decreased their speed to make it easier for the player to jump on (In the boss level). Moved the collision box.  Changed the background from pixeled to sliced.  Made it look cleaner